requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=py_0
argh=0.26.2=py38_0
astroid=2.4.2=py38_0
atomicwrites=1.4.0=py_0
attrs=20.1.0=py_0
autopep8=1.5.4=py_0
babel=2.8.0=py_0
backcall=0.2.0=py_0
bcrypt=3.2.0=py38he774522_0
blas=1.0=mkl
bleach=3.1.5=py_0
brotlipy=0.7.0=py38he774522_1000
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
cffi=1.14.2=py38h7a1dbc1_0
chardet=3.0.4=py38_1003
cloudpickle=1.5.0=py_0
colorama=0.4.3=py_0
comtypes=1.1.7=py38_1001
cryptography=3.0=py38h7a1dbc1_0
cycler=0.10.0=py38_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20200713=py_0
docutils=0.16=py38_1
entrypoints=0.3=py38_0
et_xmlfile=1.0.1=py_1001
flake8=3.8.3=py_0
freetype=2.10.2=hd328e21_0
future=0.18.2=py38_1
icu=58.2=ha925a31_3
idna=2.10=py_0
imagesize=1.2.0=py_0
importlib-metadata=1.7.0=py38_0
importlib_metadata=1.7.0=0
iniconfig=1.0.1=py_0
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.17.0=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
isort=5.4.2=py38_0
jdcal=1.4.1=py_0
jedi=0.17.1=py38_0
jinja2=2.11.2=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py38_0
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py38_0
keyring=21.3.1=py38_0
kiwisolver=1.2.0=py38h74a9793_0
lazy-object-proxy=1.4.3=py38he774522_0
libiconv=1.15=h1df5818_7
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libxml2=2.9.10=h464c3ec_1
libxslt=1.1.34=he774522_0
lxml=4.5.2=py38h1350720_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.1.2=py38_1
matplotlib-base=3.1.2=py38h64f37c6_1
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
more-itertools=8.4.0=py_0
nbconvert=5.6.1=py38_0
nbformat=5.0.7=py_0
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
numpydoc=1.1.0=py_0
openpyxl=3.0.5=py_0
openssl=1.1.1g=he774522_1
packaging=20.4=py_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py38_1
paramiko=2.7.1=py_0
parso=0.7.0=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pip=20.2.2=py38_0
pluggy=0.13.1=py38_0
prompt-toolkit=3.0.6=py_0
psutil=5.7.2=py38he774522_0
py=1.9.0=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pydocstyle=5.1.0=py_0
pyflakes=2.2.0=py_0
pygments=2.6.1=py_0
pylint=2.6.0=py38_0
pynacl=1.4.0=py38h62dcd97_1
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.16.0=py38he774522_0
pysocks=1.7.1=py38_0
pytest=6.0.1=py38_0
python=3.8.5=he1778fa_0
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.4=py_1
python-language-server=0.34.1=py38_0
pytz=2020.1=py_0
pywin32=227=py38he774522_1
pywin32-ctypes=0.2.0=py38_1000
pyyaml=5.3.1=py38he774522_1
pyzmq=19.0.1=py38ha925a31_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.7.2=py_0
qtconsole=4.7.6=py_0
qtpy=1.9.0=py_0
requests=2.24.0=py_0
rope=0.17.0=py_0
rtree=0.9.4=py38h21ff451_1
setuptools=49.6.0=py38_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.2.2=py_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.4=py38_0
spyder-kernels=1.9.3=py38_0
sqlite=3.33.0=h2a8f88b_0
testpath=0.4.4=py_0
toml=0.10.1=py_0
tornado=6.0.4=py38he774522_1
traitlets=4.3.3=py38_0
ujson=1.35=py38he774522_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
watchdog=0.10.3=py38_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py38_0
wincertstore=0.2=py38_0
wrapt=1.11.2=py38he774522_0
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
